Add option to run multiple games in one prefix.
When you try to run multiple games or programs with umu and want to run them in the same prefix you need to add the `PROTON_VERB=runinprefix`. This implements it. In the option there is a 'runinprefixplaceholder' (it needs a name that describes what it does) checkbox and when it's enabled it puts the `PROTON_VERB=runinprefix` after `PROTONPATH={runner}`.
I used fstring because everything before that uses fstring. I don't know if I got every place where it's need but after a short test it works correctly.
The only big issue I found was with shortcuts, because shortcuts run the `umu-run` directly, so if a shortcut was made before enableing this you can't run that game unless it was the first game that you launched, fixing this would require to completely rework how shortcuts work. This also doesn't work with 'Faugus Create Shortcut'.